pyproject.toml
[tool.poetry]
name = "cmd-chain"
version = "0.1.0"
description = "A simple LLM agent capable of performing simple tasks on the users operating system"
authors = ["mvishiu11 <jakubmuszynski2103@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
langchain = "^0.3.3"
openai = "^1.51.2"
uvicorn = "^0.32.0"
langchain-community = "^0.3.3"
langgraph = "^0.2.39"
python-dotenv = "^1.0.1"
langchain-google-genai = "^2.0.1"
duckduckgo-search = "^6.3.2"
langchain-google-vertexai = "^2.0.5"
langchain-anthropic = "^0.2.3"
langchain-openai = "^0.2.3"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
pylint = "^3.3.1"
mypy = "^1.12.0"
pre-commit = "^4.0.1"

[tool.poetry.scripts]
lint = "scripts.lint:main"
server = "cmd_chain.server:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
